+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - `multiSelect` is a newly added parameter.
   - Increase header rect by 3.0f to allow characters like 'g' and 'y' to be fully rendered.
   - Uses serialized property from the list instead of the cached serialized property taken in as the function's parameter.
+  - Added parameter to change header string.
 - `spaceA` and `spaceB` changed to `SPACE_A` and `SPACE_B` respectively.
 - `InspectorView` returns `Editor` when initializing.
 - `OnChange` method implemented for abstract editors. This method will be called whenever there is any changes made to the editor.